Added some info to the summary for clubs
There's a small problem in the css. When you click the arrow and it opens down, the div gets very wide. I am pretty sure it's because I am using margins to align things. I will try to fix it, but if someone knows how to do it, feel free to do it.